backend/pyproject.toml
[project]
name = "app"
version = "0.0.2"
description = ""
authors = [{ name = "Anderson Resende", email = "anderson@vinta.com.br" }]
requires-python = ">=3.10,<3.13"
readme = "README.md"
dependencies = [
    "fastapi[standard]>=0.115.0,<0.116",
    "asyncpg>=0.29.0,<0.30",
    "fastapi-users[sqlalchemy]>=13.0.0,<14",
    "pydantic-settings>=2.5.2,<3",
    "fastapi-mail>=1.4.1,<2",
    "google-generativeai>=0.8.5",
    "ruff>=0.5.3",
    "stripe>=12.1.0",
    "aiohttp>=3.11.18",
    "pygithub>=2.6.1",
    "instructor>=1.8.1",
    "anthropic>=0.51.0",
    "langfuse>=2.60.4",
    "vapi-server-sdk>=1.5.0",
    "requests>=2.32.3",
    "smolagents[litellm,toolkit]>=1.18.0",
]

[dependency-groups]
dev = [
    "pre-commit>=3.4.0,<4",
    "ruff>=0.5.3",
    "watchdog>=5.0.3,<6",
    "python-dotenv>=1.0.1,<2",
    "pytest>=8.3.3,<9",
    "pytest-mock>=3.14.0,<4",
    "mypy>=1.13.0,<2",
    "coveralls>=4.0.1,<5",
    "alembic>=1.14.0,<2",
    "pytest-asyncio>=0.24.0,<0.25",
]



[tool.hatch.build.targets.sdist]
include = ["commands"]

[tool.hatch.build.targets.wheel]
include = ["commands"]

[tool.ruff]
src = ["backend"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
